Rename api/src/auth.ts to authn.ts
